requirements.txt
djangorestframework
corsheaders
Pillow
stripe